Make some code in storeScript more readable. Log + break on AWS errors
TODO: Add the err parameter to the callback function's signature (Eg: storeScript(..., function(err, ...){}) ) for storeScript and to affiliated functions.
We now break when there's an error on producton (will give a 404 page with the current logic).
If not on production and there's an AWS error, we continue along after presenting a warning.